support dim function. device includes list of all supported operations.

+++ TelldusCoreWrapper/ITelldusCoreLibraryWrapper.cs
@@ -70,6 +70,6 @@ namespace TelldusCoreWrapper
         /// </summary>
         /// <param name="deviceId">The device id to dim</param>
         /// <param name="level">The level the device should dim to. This value should be 0 - 255</param>
-        void Dim(int deviceId, IntPtr level);
+        void Dim(int deviceId, int level);
     }
 }
